rename to remove redundant suffix from type files
https://coveord.atlassian.net/browse/KIT-2600
