+++ packages/headless/src/app/ssr-engine/types/common.ts
