+++ packages/headless/src/ssr.index.ts
@@ -6,6 +6,6 @@ polyfillCryptoNode();
 export type {Unsubscribe, Middleware} from '@reduxjs/toolkit';
 
 // ⚠️ NOTE: All exported SSR types, APIs should be marked as `@internal` until MVP is complete
-export type {InferHydrationResult} from './app/ssr-engine/types/hydrate-initial-state-ssr-types';
+export type {InferHydrationResult} from './app/ssr-engine/types/hydrate-initial-state';
 
 export {defineSearchEngine} from './app/ssr-engine/ssr-engine';
